+++ mailhub-potluck/mailhub-potluck.d/local/share/cook/bin/configure-clamav.sh
@@ -14,7 +14,7 @@ export PATH=/usr/local/bin:$PATH
 #SCRIPT=$(readlink -f "$0")
 #TEMPLATEPATH=$(dirname "$SCRIPT")/../templates
 
-# shellcheck disable=SC3003
+# shellcheck disable=SC3003,SC2039
 # safe(r) separator for sed
 #sep=$'\001'
 #