+++ mailhub-potluck/mailhub-potluck.d/local/share/cook/bin/configure-dovecot.sh
@@ -28,7 +28,7 @@ pw usermod dovecot -G vhost
 # change permissions
 chown -R vhost:vhost /mnt/dovecot
 
-# shellcheck disable=SC3003
+# shellcheck disable=SC3003,SC2039
 # safe(r) separator for sed
 sep=$'\001'